+++ archive/convert_json_to_df.py
@@ -4,7 +4,7 @@ and convert them into 3 dataframe. One for each year.
 
 import pandas as pd  # type: ignore
 import glob
-from geo_json_metrics.gps_metrics import create_df_from_json
+from gps_metrics import create_df_from_json
 import tqdm
 
 # gather all json files
